requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
smolagents
python-dotenv
transformers
huggingface-hub
dronekit
pymavlink
